backend/requirements.txt
# ── Framework ─────────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1

# ── Base de données PostgreSQL ─────────────────────────────────────────────────
sqlalchemy==2.0.36         # ORM principal
psycopg2-binary==2.9.10    # Driver PostgreSQL (binaire, pas besoin de libpq installé)
alembic==1.14.0            # Migrations de schéma (optionnel mais recommandé)